Initial support for signed metadata.
