+++ src/test/kotlin/eu/europa/ec/eudi/openid4vci/DefaultCredentialIssuerMetadataResolverTest.kt
@@ -159,4 +159,5 @@ private fun Map<CredentialConfigurationIdentifier, CredentialConfiguration>.jwtP
 private fun resolver(request: RequestMocker, expectSuccessOnly: Boolean = false) =
     CredentialIssuerMetadataResolver(
         mockedKtorHttpClientFactory(request, expectSuccessOnly = expectSuccessOnly),
+        IssuerMetadataPolicy.RequireUnsigned,
     )
