+++ src/test/kotlin/eu/europa/ec/eudi/openid4vci/IssuanceAuthorizationTest.kt
@@ -333,7 +333,7 @@ class IssuanceAuthorizationTest {
         ktorHttpClientFactory: KtorHttpClientFactory,
         credentialOfferStr: String,
     ): CredentialOffer {
-        return CredentialOfferRequestResolver(ktorHttpClientFactory = ktorHttpClientFactory)
+        return CredentialOfferRequestResolver(ktorHttpClientFactory, IssuerMetadataPolicy.RequireUnsigned)
             .resolve("https://$CREDENTIAL_ISSUER_PUBLIC_URL/credentialoffer?credential_offer=$credentialOfferStr")
             .getOrThrow()
     }
