+++ src/test/kotlin/eu/europa/ec/eudi/openid4vci/examples/PidDevIssuer.kt
@@ -40,6 +40,7 @@ internal object PidDevIssuer :
         authorizeIssuanceConfig = AuthorizeIssuanceConfig.FAVOR_SCOPES,
         parUsage = ParUsage.IfSupported,
         dPoPSigner = CryptoGenerator.ecProofSigner(),
+        issuerMetadataPolicy = IssuerMetadataPolicy.RequireUnsigned,
     )
 
     val PID_SdJwtVC_config_id = CredentialConfigurationIdentifier("eu.europa.ec.eudi.pid_vc_sd_jwt")
